Dataset federation examples
This is the section we discussed to replace the comprehensive data resources inventory, and which goes into detail about what food assistance and mental health data combinations might look like. This section has the least obvious placement for me, so if you can think of another place to put it, or another way to put this information together in a more useful way, I'm open to suggestions.